# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AutoKindler API
  version: 1.0.0
  description: API for managing arXiv document deliveries to Kindle devices.
servers:
  - url: /api
    description: Base API path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained via GitHub OAuth onboarding.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Human readable error message."
    DeliveryRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: Must be a valid arxiv.org/html/* endpoint or a .pdf link.
          example: "https://arxiv.org/html/2401.12345v1"
    Preferences:
      type: object
      properties:
        max_papers_per_month:
          type: integer
          example: 10
        category_scores:
          type: object
          additionalProperties:
            type: integer
          example:
            cs.AI: 10
            cs.LG: 8
security:
  - bearerAuth: []
paths:
  /deliveries:
    post:
      summary: Queue a new document for delivery
      description: Submits an arXiv HTML or PDF URL to be processed and sent to the user's Kindle. Rate limited to 5 per hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryRequest"
      responses:
        "202":
          description: Delivery queued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      delivery_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        example: "Pending"
        "400":
          description: Bad request (invalid URL format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "409":
          description: Conflict (Paper already delivered to user)
        "429":
          description: Too Many Requests (Rate limit exceeded)

  /deliveries/{id}:
    get:
      summary: Get delivery status
      description: Poll this endpoint to check if a delivery is Pending, Completed, or Failed.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current delivery status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      delivery_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [Pending, Completed, Failed]
                      error_message:
                        type: string
                        nullable: true
                      updated_at:
                        type: string
                        format: date-time
        "401":
          description: Unauthorized
        "404":
          description: Delivery not found

  /users/me:
    get:
      summary: Get current user profile
      description: Retrieves the authenticated user's details and subscription preferences.
      responses:
        "200":
          description: User profile object
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      kindle_email:
                        type: string
                        format: email
                      preferences:
                        $ref: "#/components/schemas/Preferences"
        "401":
          description: Unauthorized

    put:
      summary: Update user preferences
      description: Updates the user's Kindle email or daily/monthly paper preferences.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kindle_email:
                  type: string
                  format: email
                preferences:
                  $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Preferences updated successfully."
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized